+++ app/Http/Controllers/Admin/IDBInvestController.php
@@ -62,7 +62,8 @@ class IDBInvestController extends Controller
 
     public function areas_add()
     {
-        $this->authorize('resource_view');
+        $this->authorize('resource_create');
+
         return view('admin.areas.create');
     }
 
@@ -112,12 +113,16 @@ class IDBInvestController extends Controller
     }
 
     public function areas_edit($id){
+        $this->authorize('resource_edit');
+
         $data = Business::find($id); 
 
         return view('admin.areas.edit', compact('data'));
     }
 
     public function areas_update(Request $request, $id){
+        $this->authorize('resource_edit');
+
         $this->validate($request,[
             'title'     => 'required',
             'content'   => 'required',
@@ -151,9 +156,12 @@ class IDBInvestController extends Controller
     // }
     public function idb_investment_add()
     {
+        $this->authorize('resource_create');
+
         return view('admin.idb.create');
     }
     public function idb_investment_store(Request $request){
+        $this->authorize('resource_create');
 
         $this->validate($request,[
             'title'     => 'required',
@@ -177,6 +185,8 @@ class IDBInvestController extends Controller
     }
 
     public function idb_investment_status(Request $request){
+        $this->authorize('resource_status_edit');
+
         $user = Business::find($request->lid);
         $status = $request->lstatus == 1 ? '0' : '1';
         DB::transaction(function () use ($user, $status) {
@@ -189,6 +199,7 @@ class IDBInvestController extends Controller
         return response()->json($data, 200);
     }
     public function idb_investment_destroy(Request $request){
+        $this->authorize('resource_delete');
 
         $user = Business::find($request->lid);
         $user->delete();
@@ -198,12 +209,16 @@ class IDBInvestController extends Controller
     }
 
     public function idb_investment_edit($id){
+        $this->authorize('resource_edit');
+
         $data = Business::find($id); 
 
         return view('admin.idb.edit', compact('data'));
     }
 
     public function idb_investment_update(Request $request, $id){
+        $this->authorize('resource_edit');
+        
         $this->validate($request,[
             'title'     => 'required',
             'content'   => 'required',
