+++ app/Http/Controllers/ProcurementController.php
@@ -61,10 +61,11 @@ class ProcurementController extends Controller
     // }
     public function methods_add()
     {
+        $this->authorize('resource_crate');
         return view('admin.methods.create');
     }
     public function methods_store(Request $request){
-
+        $this->authorize('resource_crate');
         $this->validate($request,[
             'title'     => 'required',
             'content'   => 'required',
@@ -87,6 +88,7 @@ class ProcurementController extends Controller
     }
 
     public function methods_status(Request $request){
+        $this->authorize('resource_status_edit');
         $user = Business::find($request->lid);
         $status = $request->ustatus == 1 ? '0' : '1';
         DB::transaction(function () use ($user, $status) {
@@ -99,6 +101,7 @@ class ProcurementController extends Controller
         return response()->json($data, 200);
     }
     public function methods_destroy(Request $request){
+        $this->authorize('resource_delete');
         $user = Business::find($request->lid);
         $user->delete();
         $data['error'] = false; 
@@ -107,6 +110,7 @@ class ProcurementController extends Controller
     }
 
     public function methods_edit($id){
+        $this->authorize('resource_edit');
         $data = Business::find($id); 
         return view('admin.methods.edit', compact('data'));
     }
@@ -142,10 +146,11 @@ class ProcurementController extends Controller
 
     public function methods_services_add()
     {
+        $this->authorize('resource_create');
         return view('admin.serves.create');
     }
     public function methods_services_store(Request $request){
-
+        $this->authorize('resource_create');
         $this->validate($request,[
             'title'     => 'required',
             'content'   => 'required',
@@ -168,6 +173,7 @@ class ProcurementController extends Controller
     }
 
     public function methods_services_status(Request $request){
+        $this->authorize('resource_status_edit');
         $user = Business::find($request->lid);
         $status = $request->ustatus == 1 ? '0' : '1';
         DB::transaction(function () use ($user, $status) {
@@ -180,7 +186,7 @@ class ProcurementController extends Controller
         return response()->json($data, 200);
     }
     public function methods_services_destroy(Request $request){
-
+        $this->authorize('resource_delete');
         // dd($request->all());
         $user = Business::find($request->lid);
         $user->delete();
@@ -190,13 +196,14 @@ class ProcurementController extends Controller
     }
 
     public function methods_services_edit($id){
+        $this->authorize('resource_edit');
         $data = Business::find($id); 
 
         return view('admin.serves.edit', compact('data'));
     }
 
     public function methods_services_update(Request $request, $id){
-
+        $this->authorize('resource_edit');
         $this->validate($request,[
             'title'     => 'required',
             'content'   => 'required',
